requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.3.0  
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
loguru==0.7.2